* Reorganize the code
